Commit a sample php backend.
Currently the backend isn't working properly, it needs at least a
working db connection.

This is a reference implementation of the gpgAuth backend however. It is
not that sophisticated and improving this reference implementation would
be a great starting point. More info at http://www.gpgauth.org/

All other backends will be based off of this one. Note the README has
been updated with more information and dependencies.

No developer configuration should be necessary, I have shipped all of
the config files in this repository.