Alternative run-backend script using printf
I'm having issues with the ANSI excape sequences in the original
script. I think it has something to do with the builtin echo command
in macOSX. I rewrote the script replacing echo with printf, and
the colors work fine. I'm not sure what the convention is on cross
platform shell script compatability. If the printf works on other
systems, then maybe we should just use that. Let me know what you
think.